Added hover effect to grid
Implemented a function called changeBackground that will set background-color
of a grid square to black when a mousehover event is triggered.

Added a border to the grid as well as placeholders for the buttons to
reset and change the number of grid squares. Also made the grid container
larger to be able to fit more grid squares.

Added comments to better explain code, as well as refactored to make it easier
to understand.