Finished project
Implemented a reset button that clears out the grid and rebuilds it.
The function resetGrid() promts the user to enter a number between 8 - 100
and uses that as the "length" of each side of the grid. So entering 24
would create a grid filled with 24 cells by 24 cells. If anything outside
of the given range is entered, a default 16x16 grid is created.

Polished up the look of the app, adding an explaination to the page for the
use to quickly understand what's going on.

Project is now completed.